+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.0.1 - In progress
 ### Added
 - Validate structure of Akoya CODEX submissions.
+- Generate submission template.